CHROMIUM: config: enable ARM errata 751472
This errata claims to be relevant to all Tegra Cortex A9 systems
prior to r3p0.  We haven't necessarily verified that the problem
that the errata is addressing hits us, but we know that it is
relevant to the tegra ARM code and the hope is that this will
give us a tad bit more stability.

BUG=chromium-os:23418
TEST=Build and boot kaen

Change-Id: I1e100169a91aeb246cb46820ad6fc9683c609942
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/13990