CHROMIUM: config: enable ARM errata 720789
This errata claims to be relevant to all Tegra Cortex A9 systems
prior to r2p0.  We haven't necessarily verified that the problem
that the errata is addressing hits us, but we know that it is
relevant to the tegra ARM code and the hope is that this will
give us a tad bit more stability.

There were some claims on the mailing list thinking that this
errata wasn't needed because it was already fixed in gcc, but
the cited reference appears to be talking about something else
(TLS vs. TLB), so we're going to go ahead and turn this on.

BUG=chromium-os:23418
TEST=Build and boot kaen

Change-Id: Id65006e886886c65699fd5673105f5b7b00fc99b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/14098
Reviewed-by: Olof Johansson <olofj@chromium.org>