Revert "CHROMIUM: x86: chromeos_laptop: Enable I2C_CLIENT_WAKE for cyapa"
This reverts commit 80d4b54b437816172cb4fed301ac5fea3aac8087.

I'm reverting this change for now because the cyapa driver tries to disable
and enable the irq wake on every resume and suspend.
Because of the way that Lumpy is designed, the irq wake settings don't
actually change the wakeup settings on IRQ 21, and cause this warning.

I'll reintroduce this change once I've changed the cyapa driver to actually
use trackpad power modes to implement the irq wake/no-wake functionality.

BUG=chrome-os-partner:7652
TEST=build for lumpy. Ensure that on suspend/resume UNBALANCED
IRQ 21 wake disable warning does not occur.

Change-Id: I414441453ca09745bcf0038581fadba8311cd408
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/14655
Reviewed-by: Olof Johansson <olofj@chromium.org>