CHROMIUM: ALSA: hda - Fix calling auto_input for Cirrus codecs.
If cs421x_auto_input was called twice (like it is during init) while
the mic is present, it will over-write the last_input the second time,
causing it to switch back to the automic input when the mic is
unplugged.

Check that the mic hasn't already been switched to before setting
last_input.

BUG=chromeos-partner:7556
TEST=Boot lumpy with mic attached, record audio, remove mic, record
from internal mic, listen to recorded audio.

Change-Id: I2b249609ef2006977d99dfecb9eebe101ac6ef24
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/15098