video:msm: Move histogram reads to workqueue
Histogram must be read after each histogram completion in order to clear
the histogram registers; this patch reorganizes the histogram code to use
a deferred workqueue task to read the histogram registers out of the
interrupt context.

As well, this re-organizes the histogram start-up to clear, using the same
new workqueue, the first histogram that is read because it is often
incorrect (too many or too few bins filled).

CRs-Fixed: 328214
Change-Id: I4088028fc805113e74b5b763c6458183e75421cf
Signed-off-by: Carl Vanderlip <carlv@codeaurora.org>

Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>