msm: Disable use of the IOMMU
Originally, if CONFIG_MSM_IOMMU was enabled, clients were
expected to use the iommu for memory mapping. Currently,
it is necessary for CONFIG_MSM_IOMMU to be enabled without
having all clients ready to use the iommu. Kill support for
APIs to use the IOMMU and attaching to all domains until
clients are ready to actually use them.

Change-Id: I6ffb43f43eaa0e3e6b5a43f01b31d83d3effd970
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>