CHROMIUM: drm/i915: Fix single msg gmbus_xfers writes
gmbus_xfer with a single message (particularly a single message write) would
set Bus Cycle Select to 100b, the Gen Stop cycle, instead of 101b,
No Index, Stop cycle. This would not start single message i2c transactions.

Also, gmbus_xfer done: will disable the interface without checking if
it is idle. In the case of writes, there will be no wait on status or delay
to ensure the write starts and completes before the interface is turned off.

Fixed the former issue by using the same cycle selection as used in the
I2C_M_RD as for the write case.
GMBUS_CYCLE_WAIT | (i + 1 == num ? GMBUS_CYCLE_STOP : 0)
Fixed the latter by waiting on GMBUS_ACTIVE to deassert before disable.

BUG=chromium-os:26118
TEST=Remove the forced bitbanging in setup_gmbus, and try to write
to an i2c device on one of the GMBuses. Ensure that the transaction actually
hits the bus with a logic analyzer.

Change-Id: I36dec597314f3a43ba035cbb90994580009bbd38
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/15570
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>