CHROMUMOS: ath9k: detect rx stuck by cycle counters
And recover chip from the rx stuck.

This is a workaround from Atheros for a receive hang in their part.
It uses various status monitoring to detect that the receiver has
failed, and resets the chip.  There is activity still afoot to find
the root cause to this problem, and this change is only meant to
alleviate the immediate problem, since throughput still suffers
before the part enters this state.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:4886
TEST=Manual: Bench-top re-enactment of congested environment + in-field
testing with client machines

Change-Id: I2a742c3b7ae927b169f753a3abf8e670ebb18544
Reviewed-on: https://gerrit.chromium.org/gerrit/16731
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>