UPSTREAM: x86, efi: Add basic error handling
It's not perfect, but way better than before. Mark efi_enabled as false in
case of error and at least stop dereferencing pointers that are known to
be invalid.

The only significant missing piece is the lack of undoing the
memblock_reserve of the memory that efi marks as in use. On the other
hand, it's not a large amount of memory, and leaving it unavailable for
system use should be the safer choice anyway.

Signed-off-by: Olof Johansson <olof@lixom.net>
Link: http://lkml.kernel.org/r/1329081869-20779-5-git-send-email-olof@lixom.net
Acked-by: Matt Fleming <matt.fleming@intel.com>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>
(cherry picked from commit 140bf275d3e89e9b36851d5cf498dbbbecdf7ca8)
(staged for 3.4 at git://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git)
Signed-off-by: Anush Elangovan <anush@chromium.org>

Change-Id: Ide15b6be62320a468a194b0c2b9642c40101f42a