CHROMIUM: r8169: transition to D0 state at shutdown
With runtime PM, if the ethernet cable is disconnected, the device is
transitioned to D3 state to conserve energy. If the system is shutdown
in this state, any register accesses in rtl_shutdown are dropped on
the floor.

This patch transitions the device back to D0 state in rtl_shutdown if
the device was runtime PM suspended. In addition, since runtime PM
also mucks around with WOL options, the saved WOL options are also
restored in rtl_shutdown.

BUG=chrome-os-partner:7026
TEST=Don't have a good black box test case. It can be tested by
modifying the driver to verify that register accesses in
rtl_shutdown go through with the cable disconnected.

Change-Id: I63c843cea2ce0860f53360d5e1665ace56325f19
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/15858
Reviewed-by: Olof Johansson <olofj@chromium.org>