CHROMIUM: make perf distinguish process rename and exec. (TEMPORARY FIX.)
This makes it possible for perf report to distinguish between an exec and
a rename (for instance from prctl(PR_SET_NAME)).  Otherwise perf report
assumes all COMM records are execs and discards the old mappings even though
they may be still valid.

A larger but cleaner patch will be applied upstream (3.3. staging) so we
need to end-of-life this patch when we upgrade to that.  I could attempt
to backport that change, but it's not ready yet and it's going to be a
lot more work to maintain.

BUG=chromium-os:26115
TEST=tested with small program that uses prctl.  Testing team: no test required

Change-Id: I4fb6aadf343be07c45344c3680c4803b2dacb53d
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/15581