msm: dma: Fix the error checking for DMA clocks.
If sometime while enabling dma clocks, clock operation fails, since we
are not sanity checking the error value and assumes that dma clocks are on.
Next time when dma timer expires, it tries to disable the dma clocks which
are not enabled earlier and we get a clock unbalance stack trace warnings.

This scenario is reproducible if we continuously reboot the target. If
clk enable is failed and we bail out immediately, the next timer event
will not be set and there will be no further call to clock disable.

CRs-Fixed: 342519
Signed-off-by: Pankaj Kumar <pakuma@codeaurora.org>
(cherry picked from commit 253bafc36d5b90d1a2d8ba223f3a43ffabe1c4c7)

Change-Id: I4444e8188911051a326ad623078691f7a91abc3a
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>