CHROMIUM: drm/i915/intel_i2c: cleanup gmbus/gpio pin assignments
There is no "disabled" port 0.  So, don't even try to initialize/scan
it, etc.  This saves a bit of time when initializing the driver, since
it avoids a 50ms timeout waiting for a device to respond on a port that
doesn't even exist.

Similarly, don't initialize the "reserved" port 7, either.

Tested on Sandybridge (gen 6, PCH == CougarPoint) hardware.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:26118
TEST=no DRM error messages about "disabled" or "reserved" ports in
/var/log/messages

Change-Id: I4a0a35d9d7e356081dbcc72c7ccdf8ee3be9a7d1
Reviewed-on: https://gerrit.chromium.org/gerrit/17931
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>