seccomp_filter: pass in enter_event at creation time
[testing still in progress so no rush to review this unless you want to test it :]

With the 3.2 changes, the ftrace code stopped relying on the event id
and started relying on the passed in tp_event member value to determine
the fields for filtering.

This change replaces tp_event with a real value from where it was
uninitialized memory.  Additionally, perf_event is initialized to 0s
to help a NULL deref more quickly point out the errors.

BUG=chromium-os:27824
TEST=build amd64-generic; install; open a sample zip file; rejoice
     autotest results
  INFO    : Test results:
  ---------------------------------------------------------------
  security_Minijail_seccomp/security_Minijail_seccomp [  PASSED  ]
  ---------------------------------------------------------------
  Total PASS: 1/1 (100%)

Change-Id: Icfa68ce8d96d45b3ac35cc53a39f5a22851b00a4
Signed-off-by: Will Drewry <wad@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18295