ath9k: reduce skb truesize by 50%
Commit 0d95521ea74735826cb2e28bebf6a07392c75bfa (ath9k: use split rx
buffers to get rid of order-1 skb allocations) did a good job avoiding
using kmalloc-8192 slabs for skbs used in receive handler, and ath9k now
uses kmalloc-4096 slabs. This was done probably because of performance
issues : order-1 pages are more difficult to alloc than order-0 ones.

Problem is that Internet frames are under the 1500 bytes size, and
kmalloc-4096 is still suboptimal to hold such frames. Many drivers uses
kmalloc-2048 instead and get better performance for following reasons :

1) Permits to open receiver window faster
------------------------------------------

Sockets have a sk_rcvbuf limit guarding the total amount of memory used
by skbs in receive queue (for TCP, this includes the out of Order queue
when SACK is enabled). For each skb, the amount of memory is called
the 'truesize'. For ath9k, skbs truesize is ksize(skb->head) +
sizeof(struct sk_buff). Thats 4096 + 256 = 4352 bytes for each received
frame.

Linux opens the receive window of a tcp socket each time a new packet is
received, depending on if its a 'good citizen' or not, ie depending on
the ratio skb->len / skb->truesize.

So if a driver provides too fat skbs to tcp stack, tcp receiver window
wont grow as fast as it could be, slowing down overall download speed,
especially for long distance communications (RTT above 100 ms)

This is why its critical to make sure that regular 1500 bytes frame can
use kmalloc-2048 slab. skb->truesize being 2304 bytes instead of 4352

Some drivers make the choice to make a copy of each incoming frame to a
skb of exact size. But this adds a copy in receive path, and its not as
efficient as a zero-copy path. For USB or wireless adapters this is
probably not a problem because bandwidth is low.

2) avoids consuming too much kernel memory
-------------------------------------------
TCP maintains an out of order queue (if SACK are enabled), and this
queue can be pretty big. When/if socket sk_rcvbuf is hit, tcp stack performs
emergency 'collapsing' which is the most expensive function we have in network
stacks : It scan the whole 'out of order' queue and coalesce all frames
into new skbs to better use memory (no wasted space) and hopefully free
some memory. This blows cpu caches and is pretty slow.

Delivering skbs of 2304 bytes instead of 4352 bytes can avoid this
expensive emergency operation most of the time.

Note : I submitted an upstream patch to have kind of an fallback in tcp
stack to work around drivers that cant provide smalls skbs for any
reason :  tcp: reduce out_of_order memory use

Ref: http://patchwork.ozlabs.org/patch/147429/

BUG=chromium-os:27825
TEST=booted the machine, browsed internet, check netstat -s counters (pruned/collapsed)

Change-Id: Ic55e0d5db981e9aa98e44ac207cc6a321098beab
Signed-off-by: Eric Dumazet <edumazet@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18412
Reviewed-by: Sam Leffler <sleffler@chromium.org>