CHROMIUM: platform/x86/chromeos_laptop - refactor i2c adapter find
* Refactor find_i801 to find adapters other than SMBus I801 adapter.
This should make it easier to instantiate devices on other named
i2c adapters on the system such as i915 gmbus vga and panel without
duplicating a lot of the common name matching code.

* Clean up chromeos_laptop_dmi_table to coalesce Mario/ZGB and Link/Lumpy.
These boards use the same callback, so this is functionally the same.

* Change .ident in chromeos_laptop_dmi_table to reflect the device name
and not the board name.

BUG=chromium-os:27714,chrome-os-partner:8547
TEST=builds clean,
Test that Lumpy trackpad, cyapa on smbus, continues to be functional
on boot.

Change-Id: Icfe181c1e4e8034759ed557aaca9f7cb3e1f8bad
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18452
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18840
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>