CHROMIUM: platform/x86/chromeos_laptop - Use i2c_new_probed_device
Use i2c_new_probed_device instead of i2c_new_device. This allows
for graceful failure when a particular device is not attached to
the system, or when we are dual sourcing. We can try to add one device
and then move on to the next one if it is not probed.

BUG=chromium-os:28126
TEST=On Lumpy or Link, power down the system,
detach the camera board (which contains the als).
Boot the system.
Verify these messages do not show up in /var/log/messages
isl29018 12-0044: Write to device fails status fffffffa

Change-Id: I8a98ba0d75a2e3eba0f6a31883db46a552720cef
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18584
Reviewed-on: https://gerrit.chromium.org/gerrit/18854
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>