CHROMIUMOS: mac80211: Fix merge error involving aborting scans
A portion of the original path was not applied.  Fixed now.

BUG=chromium-os:25392
TEST=Manual: Run "iw event -f" in the background; run
"ping -i .25 <host>".  Wait for wpa_supplicant to issue a bgscan and
note that the scan is aborted due to the ping.  Issue
"iw dev wlan0 scan" and note the scan is not aborted and the ping
packets are noticeably delayed.

Change-Id: I813935dac9e563a8c76a60d3f32f8d995dfccdfd
Signed-off-by: Gary Morain <gmorain@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18979
Reviewed-by: Paul Stewart <pstew@chromium.org>