Temp change to allow upgrading the Cypress trackpad firmware
This change is necessary to force the build process to include the new
trackpad firmware in the image. It is needed only once to upgrade the
trackpad firmware, once the firmware is upgraded this change could be
removed.

The firmware blob comes from the Cypress provided file
CYTRA_115001_00_V1.0_IIC.bin.

BUG=none
TEST=manual
  . executing `echo 1 > /sys/bus/i2c/devices/2-0067/update_fw' causes
    the new firmware to be programmed into the trackpad

Change-Id: I19b75700c93b06be8e4e8f8a5d191020d29502c5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19027
Reviewed-by: Jon Salz <jsalz@chromium.org>