Subject: CHROMIUM: Input: cyapa - always try probing SMBUS if  available & i2c fails
If i2c probe fails, and the i2c adapter supports SMBUS transactions,
probe a second time, temporarily using an smbus command.

This is a one off change to work around a situation where trackpad
devices with the wrong firmware had to be updated in system.

Also:
 * force bl_enter to send reset command even if we are not sure that
we are in operational mode.
 * add sysfs entry for entering bootloader.

BUG=none
TEST=manual
  . on a system with a Trackpad with wrong firmware:
   . bring up ChromOS
   . open a shell session and execute
   . echo 1 > /sys/bus/i2c/devices/2-0067/update_fw
   . restart the system and observe trackpad operational

Change-Id: Ib261af7849ef07c1a0505d5424cdd1d7783a192d
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19028
Reviewed-by: Jon Salz <jsalz@chromium.org>