HACK: Suppress cyapa driver debug messages after successful update.
This is a hack which allows to integrate the code allowing to
reprogram the Cypress trackpad, and still avoid flooding dmesg log
with i2c related output.

The problem is that the firmware updates fail unless debug mode is
enabled. Thre is no time to investigate this, as the new devices'
trackpads need to be reimaged in the factory bringup flow, which is
almost now.

So, this hack adds a run time check, suppressing debug output in case
the trackpad s running valid firmware.

BUG=chrome-os-partner:none
TEST=manual
 . load the new image on the target
 . in bash execute the command
  echo 1 > /sys/devices/pci0000:00/0000:00:02.0/i2c-2/2-0067/update_fw
 . reboot the machine
 . check that dmesg output is not flooded with i2c messages
 . verify that the trackpad is operational

Change-Id: I8d8408f0f764a1ee78cf747097d2d6998f88dcd5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19151
Reviewed-by: Jon Salz <jsalz@chromium.org>