CHROMIUM: qcusbnet: don't spew "failed to autoresume" by default
In certain situations, we can get into a loop where we repeatedly
try to wake up the interface to send a packet.  Demote the "failed
to autoresume interface" error message to warning level so it does
not flood the logs in this case by default.

BUG=chromium-os:24525
TEST=none, can't repro on demand

Signed-off-by: ttuttle <ttuttle@chromium.org>

Change-Id: I1d1b4e2e61c125ea25c97ec9c460f2309a6b35fd
Reviewed-on: https://gerrit.chromium.org/gerrit/19245
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>