msm: gpio-v2: Fix spurious interrupts when gpio-irq is unmasked
To unmask a gpio interrupt, the gpio driver sets the
INTR_RAW_STATUS_EN and INTR_ENABLE bits in the INTR_CFG register.
As soon as the INTR_RAW_STATUS_EN is set, the INTR_STATUS is updated
and this causes a spurious interrupt when the irq is enabled.
This is noticed frequently when the gpio is pulled high and requests
a rising edge interrupt or pulled low and is requesting a falling edge
interrupt.
This causes a spurious interrupt when the irq was initialized and
during every resume. This is due the internal circuit design
of the TLMM IRQ block.
Hence to avoid the spurious irq, we set the INTR_RAW_STATUS_EN, clear the
INTR_STATUS and the set the INTR_EN to enable the interrupt.

CRs-Fixed: 346861
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
(cherry picked from commit fa15dfc0e686e62009638848308a2a69fbdd2726)

Change-Id: I2d08729a109ef6c6da45a98942db3d94df5f2587
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>