CHROMIUMOS: ath9k: fix regression in chip stuck detection
To recover the chip from stuck, the signature should be matched
for 10 interations in the 15 iteration loop. But the max values
are swapped i.e the current code will check for 15 matching
interation where the max number of iterations itself is 10. So
that it never triggers the chip reset.

The iteration check is applicable only for rxclear stuck. In
case of DCU stuck state, the reset will be triggered immediately.

BUG=chrome-os-partner:8702
TEST=manual.  Built a kernel and ran it on Alex and did some basic
WiFi connectivity.

Signed-off-by: Gary Morain <gmorain@chromium.org>
Signed-off-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19358
(cherry picked from commit 5594b4818520b740a831c75b3b74c2bf557d2839)

Change-Id: If88a4eeb7b34aeeb9b229c870a5e29de3f927758
Reviewed-on: https://gerrit.chromium.org/gerrit/19549
Reviewed-by: Gary Morain <gmorain@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>