CHROMIUMOS: ath9k: fix regression in chip stuck detection
To recover the chip from stuck, the signature should be matched
for 10 interations in the 15 iteration loop. But the max values
are swapped i.e the current code will check for 15 matching
interation where the max number of iterations itself is 10. So
that it never triggers the chip reset.

The iteration check is applicable only for rxclear stuck. In
case of DCU stuck state, the reset will be triggered immediately.

BUG=chrome-os-partner:8702
TEST=manual.  Built a kernel and ran it on Alex and did some basic
WiFi connectivity.

Change-Id: I92b2c06a70b07a3af9d1faf72b0843ebeea7650d
Signed-off-by: Gary Morain <gmorain@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19360
Reviewed-by: Paul Stewart <pstew@chromium.org>