CHROMIUM: thermal: Add a small delay before turning fans on
To prevent fans turning on due to brief spikes, if a normal fan
would be turned on schedule a second check shortly afterwards and
only turn on if the temperature remains above the threshold.

The second check is controlled by
/sys/class/thermal/thermal_zone%d/fan_on_delay which sets
the amount of time (in seconds) to delay.  A value of 0 indicates
to turn the fan on immediately.  2 seconds is the default.

Only apply this change to the normal fans, and not critical ones.

BUG=chrome-os-partner:8456
TEST=With fan_on_delay=0, run temperature up and see fans turn
    on the first poll period crossing the threshold.
  Set fan_on_delay to 2, see second checks scheduled and fans
    turn on 2 seconds temperature.
  Set fan_on_delay to 40, see fans turn on two cycles after turning on.
    Lower temperature after second check is scheduled, observe fans
      staying off.
    Lower fans under and raise above after one check has passed. See
      delay reset to 40 seconds.

Change-Id: I93afba62926ebe2359f8fd28b0ffa90417e61ba8
Signed-off-by: Jonathan Kliegman <kliegs@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19710
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>