UPSTREAM: init: fix bug where environment vars can't be passed via boot args
Commit 026cee0086f had the side-effect of dropping the '=' from
the unknown boot arguments that are passed to init as environment
variables.  This is because parse_args() puts a NUL in the string
where the '=' was when it passes the "param" and "val" pointers
to the parsing subfunctions.  Previously, unknown_bootoption() was
the last parse_args() subfunction to run, and it carefully put back
the '=' character.  Now the ignore_unknown_bootoption() is the last
one to run, and it wasn't doing the necessary repair, so the
envp params ended up with the embedded NUL and were no longer
seen as valid environment variables by init.

Tested-by: Woody Suwalski <terraluna977@gmail.com>
Acked-by: Pawel Moll <pawel.moll@arm.com>
Signed-off-by: Chris Metcalf <cmetcalf@tilera.com>
(cherry picked from commit a99cd1125189aaf31a7ee505d6208143482119eb)

BUG=chrome-os-partner:9146
TEST=boot on Link and see the dm-verity kernel command line parameter is
no longer corrupted and verified boot functions properly.

Change-Id: I017413e0ea3ddc07d2cf1a8582ac181c5164e139
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21505
Reviewed-by: Olof Johansson <olofj@chromium.org>