msm: platsmp-8625: Clear the IPC1 pending interrupt in the CORE1 context
Once core0 comes out of the PC, it uses the IPC1 interrupt to wake up
the core1 out of the sleep(GDFS). Since this interrupt is intended to
core1, move the clearing of this interrupt from core0 to core1 context.
Remove the masking and umnaking of IPC1, as no other driver is going
to use this interupt.

Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>
(cherry picked from commit a30aad0408e06249f423524b4af4b3f586ca448f)

Change-Id: Ia948becca85cbd5586e3c4e1045bab62aa06dfca
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>